Factorise le téléchargement de page en pdf
