+++ public/homologation/syntheseSecurite.js
@@ -1,23 +1,5 @@
+import telechargementPdf from '../modules/interactions/telechargementPdf.js';
+
 $(() => {
-  const telecharger = () => {
-    const doc = new jspdf.jsPDF('p', 'mm', 'a4');
-
-    const ajoutePageAvecImage = (canvas) => {
-      const img = canvas.toDataURL('image/jpeg');
-      doc.addPage();
-      doc.addImage(img, 'JPEG', 10, 10, 190, 190 * (canvas.height / canvas.width));
-    };
-
-    const construisPDF = $.map($('.page'), ($p) => $p)
-      .reduce((acc, $page) => (
-        acc.then(() => html2canvas($page, { logging: false }).then(ajoutePageAvecImage))
-      ), Promise.resolve());
-
-    construisPDF.then(() => {
-      doc.deletePage(1);
-      doc.save('syntheseSecurite.pdf');
-    });
-  };
-
-  $('#telecharger').on('click', telecharger);
+  $('#telecharger').on('click', telechargementPdf('syntheseSecurite.pdf'));
 });
