+++ src/vues/homologation/decision.pug
@@ -189,4 +189,4 @@ block page
     donnees: homologation.mesures.paginees(25)
   })
 
-  script(src='/statique/homologation/decision.js')
+  script(type = 'module', src = '/statique/homologation/decision.js')
