+++ src/vues/homologation/syntheseSecurite.pug
@@ -6,8 +6,6 @@ block page
   script(src='/bibliotheques/html2canvas.min.js')
   script(src='/bibliotheques/jspdf.umd.min.js')
 
-  include ../fragments/diagnostics
-
   block styles
     link(href='/statique/assets/styles/palette.css', rel='stylesheet')
     link(href='/statique/assets/styles/fonts.css', rel='stylesheet')
